Moved The Test Files To the __tests__ DIR
