+++ backend/src/services/issueService.ts
@@ -160,7 +160,7 @@ export default class IssueService {
 
     delete issue.issue_id;
 
-    console.log(issue);
+    // console.log(issue);
 
     const createdIssue = await this.issueRepository.createIssue({ ...issue, image_url: imageUrl });
